+++ mapwize/src/main/java/io/mapwize/mapwize/MWZMapOptions.java
@@ -15,13 +15,16 @@ public class MWZMapOptions {
     private int floor;
     private boolean zoomControl;
     private boolean showUserPositionControl = true;
+    private boolean displayFloorControl = true;
     private String accessKey;
     private String language;
     private int minZoom;
     private String outdoorMapProvider;
+    private String mainColor;
     private boolean isLocationEnabled;
     private boolean isBeaconsEnabled;
     private boolean useBrowserLocation = false;
+    private MWZCustomMarkerOptions displayMarkerOptions;
 
     public boolean isUseBrowserLocation() {
         return useBrowserLocation;
@@ -30,6 +33,30 @@ public class MWZMapOptions {
     public MWZMapOptions() {
         this.isBeaconsEnabled = true;
         this.isLocationEnabled = true;
+    }
+
+    public String getMainColor() {
+        return mainColor;
+    }
+
+    public void setMainColor(String mainColor) {
+        this.mainColor = mainColor;
+    }
+
+    public MWZCustomMarkerOptions getDisplayMarkerOptions() {
+        return displayMarkerOptions;
+    }
+
+    public void setDisplayMarkerOptions(MWZCustomMarkerOptions displayMarkerOptions) {
+        this.displayMarkerOptions = displayMarkerOptions;
+    }
+
+    public boolean isDisplayFloorControl() {
+        return displayFloorControl;
+    }
+
+    public void setDisplayFloorControl(boolean displayFloorControl) {
+        this.displayFloorControl = displayFloorControl;
     }
 
     public boolean isLocationEnabled() {